feat: Add mood event mapping functionality Closes #73, Closes #74, Closes#75
